+++ htmllaundry/utils.py
@@ -78,6 +78,10 @@ def remove_empty_tags(doc, extra_empty_tags=[]):
             if el.tag in legal_empty_tags:
                 continue
 
+            # Empty <a> can be used as anchor.
+            if (el.tag == 'a') and (('name' in el.attrib) or ('id' in el.attrib)):
+                continue
+
             if len(el) == 0 and is_whitespace(el.text):
                 victims.append(el)
                 continue